requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
mistralai>=0.1.8
python-dotenv>=1.0.0


# to install or reinstall dependencies use in GIT/Terminal: 
# pip install -r requirements.txt --upgrade --force-reinstall